navigationcontrollerに処理を移した
